feat(core): changes the way creation of object in collection work. Adds $reveal and $moveTo methods.
Creation of new objects in collection no longer adds the object inmediatelly, it now waits for server success response to call add.

Alternatively the `$reveal` can be used to add objects inmediatelly or to prevent objects to be added on success.

Closes #49

BREAKING CHANGES: New objects are not inmediatelly added to collections anymore. Also, the $load method has been removed.

To migrate existing code:

* replace `$load` calls by $new.
* replace calls to $build/$new/$buildRaw/$create by <method>().$reveal() to inmediatelly show new objects.