Added support for multiple model types to shared.js
Allows shared.js to store cache based on type of data, so it can store multiple objects with the same $pk, as long as they are of different types.  Most of this code came from @jasonayre in issue #208.
Lines  39-41 were added by me to fix a bug where model scopes would get nested after being cached.  This way the scope is refreshed every time it is read from the cache.